requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
av==10.0.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.2.0
ctranslate2==3.23.0
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faster-whisper==0.10.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
googletrans==3.0.0
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.2
langchain-community==0.0.14
langchain-core==0.1.14
langchain-openai==0.0.3
langsmith==0.0.83
lxml==5.1.0
marshmallow==3.20.2
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
onnxruntime==1.16.3
openai==1.9.0
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.25.1
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
yarl==1.9.4
